refactor `hvm.rs`
